4-5まで

+++ 3-1.php
@@ -1,6 +1,6 @@
 <?php
     echo "整数値：";
-    $return = fgets(STDIN);
+    $return = (int)fgets(STDIN);
 
     if ($return < 0)
         echo "その値は負です。\n";
